+++ pennylane_sf/__init__.py
@@ -16,5 +16,5 @@ Top level PennyLane-SF module.
 """
 from .fock import StrawberryFieldsFock
 from .gaussian import StrawberryFieldsGaussian
-from .hw import StrawberryFieldsRemote
+from .remote import StrawberryFieldsRemote
 from ._version import __version__
